style: move BigInteger extensions from Utility to BigIntegerExtensions (#3916)
