--- src/Neo.VM/Utility.cs
@@ -1,80 +1,0 @@
-// Copyright (C) 2015-2025 The Neo Project.
-//
-// Utility.cs file belongs to the neo project and is free
-// software distributed under the MIT software license, see the
-// accompanying file LICENSE in the main directory of the
-// repository or http://www.opensource.org/licenses/mit-license.php
-// for more details.
-//
-// Redistribution and use in source and binary forms with or without
-// modifications are permitted.
-
-using Neo.VM.Types;
-using System;
-using System.Numerics;
-using System.Runtime.CompilerServices;
-
-namespace Neo.VM
-{
-    public static class Utility
-    {
-        public static BigInteger Sqrt(this BigInteger value)
-        {
-            if (value < 0) throw new InvalidOperationException($"value {value} can not be negative for {nameof(Integer)}.{nameof(Sqrt)}.");
-            if (value.IsZero) return BigInteger.Zero;
-            if (value < 4) return BigInteger.One;
-
-            var z = value;
-            var x = BigInteger.One << (int)(((value - 1).GetBitLength() + 1) >> 1);
-            while (x < z)
-            {
-                z = x;
-                x = (value / x + x) / 2;
-            }
-
-            return z;
-        }
-
-        /// <summary>
-        /// Gets the number of bits required for shortest two's complement representation of the current instance without the sign bit.
-        /// </summary>
-        /// <returns>The minimum non-negative number of bits in two's complement notation without the sign bit.</returns>
-        /// <remarks>This method returns 0 if the value of current object is equal to <see cref="BigInteger.Zero"/> or <see cref="BigInteger.MinusOne"/>. For positive integers the return value is equal to the ordinary binary representation string length.</remarks>
-        public static long GetBitLength(this BigInteger value)
-        {
-#if NET5_0_OR_GREATER
-            return value.GetBitLength();
-#else
-            if (value == 0 || value == BigInteger.MinusOne) return 0;
-
-            // Note: This method is imprecise and might not work as expected with integers larger than 256 bits.
-            var b = value.ToByteArray();
-            if (b.Length == 1 || (b.Length == 2 && b[1] == 0))
-            {
-                return BitCount(value.Sign > 0 ? b[0] : (byte)(255 - b[0]));
-            }
-            return (b.Length - 1) * 8 + BitCount(value.Sign > 0 ? b[^1] : 255 - b[^1]);
-#endif
-        }
-
-#if !NET5_0_OR_GREATER
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int BitCount(int w)
-        {
-            return w < 1 << 15 ? (w < 1 << 7
-                ? (w < 1 << 3 ? (w < 1 << 1
-                ? (w < 1 << 0 ? (w < 0 ? 32 : 0) : 1)
-                : (w < 1 << 2 ? 2 : 3)) : (w < 1 << 5
-                ? (w < 1 << 4 ? 4 : 5)
-                : (w < 1 << 6 ? 6 : 7)))
-                : (w < 1 << 11
-                ? (w < 1 << 9 ? (w < 1 << 8 ? 8 : 9) : (w < 1 << 10 ? 10 : 11))
-                : (w < 1 << 13 ? (w < 1 << 12 ? 12 : 13) : (w < 1 << 14 ? 14 : 15)))) : (w < 1 << 23 ? (w < 1 << 19
-                ? (w < 1 << 17 ? (w < 1 << 16 ? 16 : 17) : (w < 1 << 18 ? 18 : 19))
-                : (w < 1 << 21 ? (w < 1 << 20 ? 20 : 21) : (w < 1 << 22 ? 22 : 23))) : (w < 1 << 27
-                ? (w < 1 << 25 ? (w < 1 << 24 ? 24 : 25) : (w < 1 << 26 ? 26 : 27))
-                : (w < 1 << 29 ? (w < 1 << 28 ? 28 : 29) : (w < 1 << 30 ? 30 : 31))));
-        }
-#endif
-    }
-}
