+++ src/Neo/Cryptography/ECC/ECCurve.cs
@@ -46,7 +46,7 @@ namespace Neo.Cryptography.ECC
         private ECCurve(BigInteger Q, BigInteger A, BigInteger B, BigInteger N, byte[] G, string curveName)
         {
             this.Q = Q;
-            ExpectedECPointLength = ((int)VM.Utility.GetBitLength(Q) + 7) / 8;
+            ExpectedECPointLength = ((int)Q.GetBitLength() + 7) / 8;
             this.A = new ECFieldElement(A, this);
             this.B = new ECFieldElement(B, this);
             this.N = N;
