+++ src/Neo/Cryptography/ECC/ECPoint.cs
@@ -234,7 +234,7 @@ namespace Neo.Cryptography.ECC
         internal static ECPoint Multiply(ECPoint p, BigInteger k)
         {
             // floor(log2(k))
-            var m = (int)VM.Utility.GetBitLength(k);
+            var m = (int)k.GetBitLength();
 
             // width of the Window NAF
             sbyte width;
@@ -397,7 +397,7 @@ namespace Neo.Cryptography.ECC
 
         private static sbyte[] WindowNaf(sbyte width, BigInteger k)
         {
-            var wnaf = new sbyte[VM.Utility.GetBitLength(k) + 1];
+            var wnaf = new sbyte[k.GetBitLength() + 1];
             var pow2wB = (short)(1 << width);
             var i = 0;
             var length = 0;
